+++ client/src/Providers/spDefaultItems.ts
@@ -1,0 +1,2 @@
+export const defaultConstantItems = ["true", "false", "null"];
+export const defaultKeywordsItems = ["case", "switch"];