+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [2.2.2]
 
+### Added
+
+- Added keywords and constants completions.
+
 ### Fixed
 
 - Fixed parsing of old syntax (SM1.7) functions.